# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kasir API",
    "version": "1.0.0",
    "description": "REST API sederhana untuk manajemen produk dan kategori (in-memory)."
  },
  "servers": [
    { "url": "https://kasir-api-seven.vercel.app" }
  ],
...
  "paths": {
...
    "/api/product": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductInput" },
              "example": { "name": "Beras", "price": 15000, "stock": 20 }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          }
        }
      }
    },
...
    "/api/category": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryInput" },
              "example": { "name": "Sembako", "description": "Produk kebutuhan sehari-hari" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "price": { "type": "integer" },
          "stock": { "type": "integer" }
        }
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "price": { "type": "integer" },
          "stock": { "type": "integer" }
        },
        "required": ["name", "price", "stock"]
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "CategoryInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" }
        },
        "required": ["name", "description"]
      },
...
    }
  }
}